+++ app/(tabs)/map.tsx
@@ -1,20 +1,5 @@
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { StyleSheet } from 'react-native';
+import MapView from '../../components/map';
 
 export default function MapScreen() {
-  return (
-    <ThemedView style={styles.container}>
-      <ThemedText type="title">Función no disponible</ThemedText>
-      <ThemedText>El mapa de recursos solo está disponible en la aplicación móvil.</ThemedText>
-    </ThemedView>
-  );
+  return <MapView />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
