+++ components/map/index.native.tsx
@@ -1,0 +1,1 @@
+export { default } from './MapView.native';
