+++ components/map/index.web.tsx
@@ -1,0 +1,1 @@
+export { default } from './MapView.web';
